Updated fixsamba to copy smb.conf AFTER killing smbd
The previous version of this script copied /jffs/smb.conf to /etc/smb.conf before killing the smbd daemon, but WRT firmware was overwriting /etc/smb.conf upon stopping the daemon. Re-ordered actions to kill the daemon, wait 3 minutes, then copy the custom smb.conf file and restart the daemon.

I'm not entirely certain the 3 minute delay is necessary, but I'm not certain when the file is being overwritten, and the 3 minute delay seems to work consistently.

Thanks to @stevland for pointing this out.